added read, update and delete notes

+++ app/src/main/java/com/notes/chan/simplenotetakingapp/DbHelper.java
@@ -61,7 +61,7 @@ public class DbHelper extends SQLiteOpenHelper {
         return id;
     }
 
-    public boolean updateNote(SQLiteDatabase db, int id, String title, String body){
+    public boolean updateNote(SQLiteDatabase db, long id, String title, String body){
         Calendar c = Calendar.getInstance();
         int year = c.get(Calendar.YEAR);
         int month = c.get(Calendar.MONTH) + 1;
@@ -76,7 +76,7 @@ public class DbHelper extends SQLiteOpenHelper {
         return numberOfRowsAffected > 0;
     }
 
-    public boolean deleteNote(SQLiteDatabase db, int id){
+    public boolean deleteNote(SQLiteDatabase db, long id){
         int numberOfRowsDeleted = db.delete(TABLE_NOTES, ID + "=" + id, null);
         return numberOfRowsDeleted > 0;
     }
